chore: move states home class to home states file
