+++ lib/src/modules/home/presentation/controller/states/home_state.dart
@@ -1,0 +1,19 @@
+import '../../../../../core/data/models/packages_model.dart';
+
+abstract interface class HomeState {}
+
+final class IdleHomeState extends HomeState {}
+
+final class LoadingHomeState extends HomeState {}
+
+final class SuccessHomeState extends HomeState {
+  List<PackagesModel> list;
+
+  SuccessHomeState({required this.list});
+}
+
+final class ErrorHomeState extends HomeState {
+  String message;
+
+  ErrorHomeState({required this.message});
+}
